Day 7/Industry_Event_Participation_Verifier/requirements.txt
streamlit
langchain
faiss-cpu
PyMuPDF
python-dotenv
google-generativeai
Pillow
langchain-google-genai
PyPDF2
langchain-community
